Show price charts on News CoinBody  (#383)
* Show price charts on News CoinBody #811229137592919

* Return a new datum object with timestamp instead of mutating priceData #811229137592919
