--- app/javascript/App/components/CoinCharts/PriceGraph/parseData.js
@@ -1,11 +1,0 @@
-export default ({ priceData }) => {
-  const prices = []
-  const volume = []
-  priceData.forEach((day) => {
-    let { timestamp: time, close: price, volume_from: vol } = day
-    prices.push([time, price])
-    volume.push([time, vol])
-    //debugger
-  })
-  return { prices, volume }
-}
