+++ app/javascript/App/components/CoinShow/index.js
@@ -67,11 +67,6 @@ class CoinShow extends Component {
       coinObj,
     } = this.props
 
-    const epochPrices = priceData.map((datum) => {
-      datum.timestamp = new Date(datum.time).getTime()
-      return datum
-    })
-
     const percentChange1h = {
       positive: coinObj.change1h > 0,
       value: coinObj.change1h,
@@ -185,7 +180,7 @@ class CoinShow extends Component {
                       <Card title="Price Chart" style={{ padding: 1 }}>
                         <CoinCharts
                           symbol={symbol}
-                          priceData={epochPrices}
+                          priceData={priceData}
                           annotations={annotations}
                           isTradingViewVisible={isTradingViewVisible}
                         />
